Change resolution-type int -> float

+++ app/project/api/schemas/configuration_schema.py
@@ -27,7 +27,7 @@ class ConfigurationSchema(Schema):
     label = fields.String(allow_none=True)
     status = fields.String(default="draft", allow_none=True)
     hierarchy = ConfigurationHierarchyField(allow_none=True)
-    resolution = fields.Integer(allow_none=True)
+    resolution = fields.Float(allow_none=True)
     resolution_unit_uri = fields.String(allow_none=True)
     resolutionUnitName = fields.String(allow_none=True)
